backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
python-decouple
psycopg2-binary